* refactored external info pages to new abstracted DOM-model paradigm
